ddr_Fastapi/requirements.txt
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==4.2.0
asyncpg==0.29.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.0
ecdsa==0.18.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.26.0
